% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R
\docType{import}
\name{pipe-operator}
\alias{pipe-operator}
\alias{\%>\%}
\title{Pipe operator}
\description{
This is the 'magrittr' pipe operator. We import it in this
package in order to help the user to construct the \code{model} argument of
the \code{\link{getAb}} function.
}
